docs: rename files in camelCase, delete no-used
