--- src/js/previousKey.js
@@ -1,2 +1,0 @@
-const previousKey = ' ';
-export default previousKey;
